Lookup table values closer to expected from indexing fix

+++ hls4ml/templates/catapult/nnet_utils/nnet_activation_stream.h
@@ -620,7 +620,7 @@ void softplus(ac_channel<data_T> &data, ac_channel<res_T> &res) {
         #pragma hls_unroll
         SoftplusPackLoop: for (int j = 0; j < res_T::size; j++) {
             //#pragma HLS UNROLL
-            int data_round = (int)in_data[j].to_double()*(int)CONFIG_T::table_size/16;
+            int data_round = in_data[j].to_double()*(int)CONFIG_T::table_size/16;
             int index = data_round + 8*(int)CONFIG_T::table_size/16;
             if (index < 0)   index = 0;
             else if (index > CONFIG_T::table_size-1) index = (int)CONFIG_T::table_size-1;
